requirements.txt
requests
python-dotenv
openai
flask
pydantic
humanize
gitpython
sentry-sdk[flask]
flask-socketio